fix: 형변환 수정

See merge request s11-webmobile1-sub2/S11P12D107!333

+++ backend/src/main/java/com/chestnut/backend/log/entity/StudyLog.java
@@ -53,7 +53,7 @@ public class StudyLog {
             Member member,
             Chapter chapter,
             Study study,
-            byte todayCount,
+            int todayCount,
             boolean isPass,
             boolean passRecord
     ) {
